refactor(server-utils): extract project root logic into separate function
